// packages/react/src/models/branding-preferences.ts
export interface BrandingPreferenceImagesInterface {
  /**
     * Organization Favicon.
     */
  favicon: Omit<BrandingPreferenceImageInterface, 'altText'>;
  /**
     * Organization Logo.
     */
  logo: BrandingPreferenceImageInterface;
  /**
     * Organization My Account Logo.
     */
  myAccountLogo: BrandingPreferenceImageInterface;
}

/**
 * Interface Branding preference image.
 */
export interface BrandingPreferenceImageInterface {
  /**
     * Image Alt.
     */
  altText: string | undefined;
  /**
     * Image URL.
     */
  imgURL: string | undefined;
  /**
     * Title.
     */
  title?: string;
}
// ...
export interface ThemeConfigInterface {
  /**
     * Button Preferences.
     */
  buttons: BrandingPreferenceButtonsInterface;
  /**
     * Color Palette.
     */
  colors: BrandingPreferenceColorsInterface;
  /**
     * Footer Preferences.
     */
  footer: BrandingPreferenceFooterInterface,
  /**
     * images such as Logo, Favicon, etc.
     */
  images: BrandingPreferenceImagesInterface;
  /**
     * Input Fields Preferences.
     */
  inputs: ElementStateInterface<BrandingPreferenceInputInterface>;
  /**
     * Login Box Preferences.
     */
  loginBox: BrandingPreferenceLoginBoxInterface;
  /**
     * Login Page Preferences.
     */
  loginPage?: BrandingPreferencePageInterface;
  /**
     * Page Preferences.
     * @deprecated Renamed to `loginPage` to keep it specific for login page.
     */
  page?: BrandingPreferencePageInterface;
  /**
     * Typography Preferences.
     */
  typography: BrandingPreferenceTypographyInterface;
}

/**
 * Strict Interface Branding preference theme.
 */
export interface StrictBrandingPreferenceThemeInterface {
  /**
     * The active theme.
     */
  activeTheme: PredefinedThemes;
}

/**
 * Represents a color palette with different shades and contrast text.
 */
export interface PaletteColor {
  /**
     * The contrast text color for the color.
     */
  contrastText?: string;

  /**
     * The dark shade of the color.
     */
  dark?: string;

  /**
     * The inverted color for the color.
     */
  inverted?: string;

  /**
     * The light shade of the color.
     */
  light?: string;

  /**
     * The main shade of the color.
     */
  main: string;
}

/**
 * Interface defining the color palette for a branding preference theme.
 */
export interface BrandingPreferenceColorsInterface {
  /**
     * The alerts color palette of the theme.
     */
  alerts: {
    /**
         * The error alerts color palette of the theme.
         */
    error: PaletteColor;
    /**
         * The info alerts color palette of the theme.
         */
    info: PaletteColor;
    /**
         * The neutral alerts color palette of the theme.
         */
    neutral: PaletteColor;
    /**
         * The warning alerts color palette of the theme.
         */
    warning: PaletteColor;
  },
  /**
     * The background color palette of the theme.
     */
  background: {
    /**
         * The body background color palette of the theme.
         */
    body: PaletteColor;
    /**
         * The surface background color palette of the theme.
         */
    surface: PaletteColor;
  };
  /**
     * The illustrations color palette of the theme.
     */
  illustrations: {
    /**
         * The accent 1 illustrations color palette of the theme.
         */
    accent1: PaletteColor;
    /**
         * The accent 2 illustrations color palette of the theme.
         */
    accent2: PaletteColor;
    /**
         * The accent 3 illustrations color palette of the theme.
         */
    accent3: PaletteColor;
    /**
         * The primary illustrations color palette of the theme.
         */
    primary: PaletteColor;
    /**
         * The secondary illustrations color palette of the theme.
         */
    secondary: PaletteColor;
  },
  /**
     * The outlined color palette of the theme.
     */
  outlined: {
    /**
         * The default outlined color palette of the theme.
         */
    default: string;
  },
  /**
     * The primary color palette of the theme.
     */
  primary: PaletteColor;
  /**
     * The secondary color palette of the theme.
     */
  secondary: PaletteColor;
  /**
     * The text color palette of the theme.
     */
  text: {
    /**
         * The primary text color palette of the theme.
         */
    primary: string;
    /**
         * The secondary text color palette of the theme.
         */
    secondary: string;
  }
}

/**
 * Interface Branding preference footer preferences.
 */
export interface BrandingPreferenceFooterInterface {
  /**
     * Page Body Font.
     */
  border: Pick<BorderStyleAttributesInterface, 'borderColor'>;
  /**
     * Page Body Font.
     */
  font: FontStyleAttributesInterface;
}

/**
 * Interface Branding preference page preferences.
 */
export interface BrandingPreferencePageInterface {
  /**
     * Page Background.
     */
  background: BackgroundStyleAttributesInterface;
  /**
     * Page Body Font.
     */
  font: FontStyleAttributesInterface;
}

/**
 * Interface for the Branding Preference Typography.
 */
export interface BrandingPreferenceTypographyInterface {
  /**
     * Page Font.
     */
  font: BrandingPreferenceTypographyFontInterface;
  /**
     * Page Heading Typography.
     */
  heading: {
    /**
         * Page Heading Font Preferences.
         */
    font: ColorStyleAttributesInterface
  };
}

/**
 * Interface for the Font Typography Font.
 */
export interface BrandingPreferenceTypographyFontInterface {
  /**
     * Font Family.
     */
  fontFamily: string;
  /**
     * URL to import if loaded from a CDN.
     */
  importURL?: string;
}

/**
 * Interface for the Login Box Preferences.
 */
export interface BrandingPreferenceButtonsInterface {
  /**
     * Social, External IDP Connection Button Preference.
     */
  externalConnection: ElementStateInterface<ButtonStyleAttributesInterface>;
  /**
     * Primary Button Preferences.
     */
  primary: ElementStateInterface<Omit<ButtonStyleAttributesInterface, 'background'>>;
  /**
     * Secondary Button Preferences.
     */
  secondary: ElementStateInterface<Omit<ButtonStyleAttributesInterface, 'background'>>;
}

/**
 * Interface for the Login Box Preferences.
 */
export interface BrandingPreferenceInputInterface {
  /**
     * Input field background.
     */
  background: BackgroundStyleAttributesInterface;
  /**
     * Secondary Button Preferences.
     */
  border: Pick<BorderStyleAttributesInterface, 'borderRadius' | 'borderColor'>;
  /**
     * Input Field Font Preferences.
     */
  font: FontStyleAttributesInterface;
  /**
     * Input Labels Preferences.
     */
  labels: {
    /**
         * Input Labels Font Preferences.
         */
    font: FontStyleAttributesInterface;
  };
}

export interface BrandingPreferenceLoginBoxInterface {
  /**
     * Login Box Background.
     */
  background: BackgroundStyleAttributesInterface;
  /**
     * Login Box Border.
     */
  border: BorderStyleAttributesInterface;
  /**
     * Login Box Font.
     */
  font: FontStyleAttributesInterface;
}
// ...
export enum PredefinedThemes {
  DARK = 'DARK',
  LIGHT = 'LIGHT',
}